+++ server/src/main/java/org/elasticsearch/index/mapper/IpFieldMapper.java
@@ -68,14 +68,15 @@ public class IpFieldMapper extends ParametrizedFieldMapper {
 
         private final Parameter<Boolean> ignoreMalformed;
         private final Parameter<InetAddress> nullValue = new Parameter<>("null_value", false, () -> null,
-            (n, c, o) -> InetAddresses.forString(o.toString()), m -> toType(m).nullValue)
+            (n, c, o) -> o == null ? null : InetAddresses.forString(o.toString()), m -> toType(m).nullValue)
             .setSerializer((b, f, v) -> {
                 if (v == null) {
                     b.nullField(f);
                 } else {
                     b.field(f, InetAddresses.toAddrString(v));
                 }
-            }, NetworkAddress::format);
+            }, NetworkAddress::format)
+            .acceptsNull();
 
         private final Parameter<Map<String, String>> meta = Parameter.metaParam();
