+++ server/src/test/java/org/elasticsearch/index/mapper/NullValueTests.java
@@ -1,16 +1,3 @@
-package org.elasticsearch.index.mapper;
-
-import org.elasticsearch.common.Strings;
-import org.elasticsearch.common.compress.CompressedXContent;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.xcontent.XContentFactory;
-import org.elasticsearch.index.IndexService;
-import org.elasticsearch.test.ESSingleNodeTestCase;
-
-import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.either;
-import static org.hamcrest.Matchers.equalTo;
-
 /*
  * Licensed to Elasticsearch under one or more contributor
  * license agreements. See the NOTICE file distributed with
@@ -30,32 +17,34 @@ import static org.hamcrest.Matchers.equalTo;
  * under the License.
  */
 
-public class NullValueTests extends ESSingleNodeTestCase {
+package org.elasticsearch.index.mapper;
+
+import org.elasticsearch.common.Strings;
+import org.elasticsearch.common.xcontent.ToXContent;
+import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.json.JsonXContent;
+
+import java.util.Collections;
+import java.util.Map;
+
+import static org.hamcrest.Matchers.containsString;
+
+public class NullValueTests extends MapperServiceTestCase {
+
     public void testNullNullValue() throws Exception {
-        IndexService indexService = createIndex("test", Settings.builder().build());
+
         String[] typesToTest = {"integer", "long", "double", "float", "short", "date", "ip", "keyword", "boolean", "byte", "geo_point"};
 
         for (String type : typesToTest) {
-            String mapping = Strings.toString(XContentFactory.jsonBuilder()
-                    .startObject()
-                        .startObject("type")
-                            .startObject("properties")
-                                .startObject("numeric")
-                                    .field("type", type)
-                                    .field("null_value", (String) null)
-                                .endObject()
-                            .endObject()
-                        .endObject()
-                    .endObject());
+            DocumentMapper mapper = createDocumentMapper(fieldMapping(b -> b.field("type", type).nullField("null_value")));
 
-            try {
-                indexService.mapperService().documentMapperParser().parse("type", new CompressedXContent(mapping));
-                fail("Test should have failed because [null_value] was null.");
-            } catch (MapperParsingException e) {
-                assertThat(e.getMessage(),
-                    either(equalTo("Property [null_value] cannot be null."))
-                    .or(containsString("must not have a [null] value")));
-            }
+            mapper.parse(source(b -> b.nullField("field")));
+
+            ToXContent.Params params = new ToXContent.MapParams(Collections.singletonMap("include_defaults", "true"));
+            XContentBuilder b = JsonXContent.contentBuilder().startObject();
+            mapper.mapping().toXContent(b, params);
+            b.endObject();
+            assertThat(Strings.toString(b), containsString("\"null_value\":null"));
         }
     }
 }
